Implemented playing a land
Additionally, allowed a card to provide name and type, implemented hand as a stratification of cards, allowed a part of the battlefield to provide its number of cards and content, allowed a permanent to provide its name, and implemented a planeswalker card and an artifact card.

Going to implement playing other types of cards.